add: separate treasury tests
